requirements.txt
asgiref==3.5.2
bandit==1.7.4
cffi==1.15.1
cryptography==38.0.0
Django==4.0
django-appconf==1.0.5
django-axes==5.39.0
django-cryptography==1.1
django-formtools==2.3
django-ipware==4.0.2
django-otp==1.1.3
django-phonenumber-field==6.4.0
django-sslserver==0.22
django-two-factor-auth==1.14.0
gitdb==4.0.9
GitPython==3.1.27
pbr==5.10.0
phonenumbers==8.12.55
pycparser==2.21
pytz==2022.2.1
PyYAML==6.0
qrcode==7.3.1
smmap==5.0.0
sqlparse==0.4.2
stevedore==4.0.0
